+++ app/[lang]/(app)/(home)/(tablet)/room-devision/rooms-management/entrance-and-exit-rooms/EntranceAndExitRoomsWrapper.tsx
@@ -1,0 +1,16 @@
+import { type EntranceAndExitRoomsDictionary } from '@/internalization/app/dictionaries/(tablet)/room-devision/entrance-and-exit-rooms/dictionary';
+import EntranceAndExitFilters from './components/EntranceAndExitFilters';
+import EntranceAndExitList from './components/EntranceAndExitList';
+
+export default function EntranceAndExitRoomsWrapper({
+ dic,
+}: {
+ dic: EntranceAndExitRoomsDictionary;
+}) {
+ return (
+  <div>
+   <EntranceAndExitFilters dic={dic} />
+   <EntranceAndExitList dic={dic} />
+  </div>
+ );
+}