+++ app/[lang]/(app)/(home)/(tablet)/room-devision/rooms-management/entrance-and-exit-rooms/components/EntranceAndExitFilters.tsx
@@ -1,0 +1,7 @@
+import { type EntranceAndExitRoomsDictionary } from '@/internalization/app/dictionaries/(tablet)/room-devision/entrance-and-exit-rooms/dictionary';
+
+export default function EntranceAndExitFilters({}: {
+ dic: EntranceAndExitRoomsDictionary;
+}) {
+ return <div></div>;
+}