+++ app/[lang]/(app)/(home)/(tablet)/room-devision/rooms-management/entrance-and-exit-rooms/components/EntranceAndExitList.tsx
@@ -1,0 +1,9 @@
+import { type EntranceAndExitRoomsDictionary } from '@/internalization/app/dictionaries/(tablet)/room-devision/entrance-and-exit-rooms/dictionary';
+
+export default function EntranceAndExitList({
+ dic,
+}: {
+ dic: EntranceAndExitRoomsDictionary;
+}) {
+ return <div></div>;
+}