+++ app/[lang]/(app)/(home)/(tablet)/room-devision/rooms-management/page.tsx
@@ -1,5 +1,6 @@
 import { getRoomsManagementDictionary } from '@/internalization/app/dictionaries/(tablet)/room-devision/rooms-management/dictionary';
 import { getRoomStatisticsDictionary } from '@/internalization/app/dictionaries/(tablet)/room-devision/room-statistics/dictionary';
+import { getEntranceAndExitRoomsDictionary } from '@/internalization/app/dictionaries/(tablet)/room-devision/entrance-and-exit-rooms/dictionary';
 import { type Locale } from '@/internalization/app/localization';
 import RoomsManagementWrapper from './components/RoomsManagementWrapper';
 
@@ -13,10 +14,14 @@ export default async function RoomsManagement(
  const roomsStatisticsDic = await getRoomStatisticsDictionary({
   locale: lang as Locale,
  });
+ const entranceAndExitDic = await getEntranceAndExitRoomsDictionary({
+  locale: lang as Locale,
+ });
  return (
   <RoomsManagementWrapper
    dic={roomsManagementDic}
    roomsStatisticsDic={roomsStatisticsDic}
+   entranceAndExitDic={entranceAndExitDic}
   />
  );
 }